Crate proj_net

Proj_Net library crate.

This library implements networking around the protocol implemented in proj_crypto.

The cryptography in proj_crypto has not been reviewed.

For example usage see the server_echo() test in lib.rs and the interactive demo in main.rs.

This project is licenced under the terms of the GNU General Public Licence as published by the Free Software Foundation, either version 3 of the licence, or (at your option) any later version.

Modules

client
Client functionality

server
Server functionality

Structs

SessionKeys
Stores session keys

Type Definitions

Keypair
Simple tuple of a public key and a secret key